test: split relay runtime coverage
